subida automatica

+++ ft_print_pointer.c
@@ -6,7 +6,7 @@
 /*   By: sescobar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/17 15:50:39 by sescobar          #+#    #+#             */
-/*   Updated: 2020/10/13 13:43:21 by sescobar         ###   ########.fr       */
+/*   Updated: 2020/10/13 13:50:44 by sescobar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,10 +47,10 @@ char		*width_p(char *hex, t_flags **flags)
 int			ft_print_pointer(const char *str, va_list args,
 		int writed, t_flags **flags)
 {
-	unsigned int	n;
+	unsigned long	n;
 	char			*hex;
 
-	n = va_arg(args, unsigned int);
+	n = va_arg(args, unsigned long);
 	if ((((*flags)->has_precission == 1 && (*flags)->precission == 0) ||
 		((*flags)->has_width == 1 && (*flags)->width == 0)) && (n == 0))
 		hex = ft_strjoin_free("", "", 0);
